// src/app/documents/[id]/types.ts
export interface DocumentChunk {
    id: string;
    sequence: number;
    content: string;
    meta_data: {
      length: number;
      index: number;
    };
  }
  
export interface DocumentMetadata {
    title: string;
    pages: number;
    text: string;
    content?: string;
    chunks: DocumentChunk[];
    meta_data: {
      chunks_count: number;
    };
  }
  
export interface DocumentMetadataResponse {
    document: DocumentMetadata;
  }